Tuning security & refactoring
