--- BusinessAssistant-mydata/src/main/java/com/businessassistantbcn/mydata/config/DisableSecurityConfig.java
@@ -1,19 +1,0 @@
-package com.businessassistantbcn.mydata.config;
-
-/*import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;*/
-
-/*@Configuration
-@EnableAutoConfiguration*/
-/*@Profile("test")*/
-public class DisableSecurityConfig{/* extends WebSecurityConfigurerAdapter {
-
-    @Override
-    public void configure(WebSecurity web) throws Exception {
-        web.ignoring().antMatchers("/**");
-    }*/
-}
