+++ BusinessAssistant-mydata/src/test/java/com/businessassistantbcn/mydata/config/DBTestConfiguration.java
@@ -1,33 +1,25 @@
 package com.businessassistantbcn.mydata.config;
 
-/*import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.annotation.Order;
 import org.springframework.core.env.Environment;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.transaction.annotation.EnableTransactionManagement;*/
+import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 import javax.sql.DataSource;
 
-/*@Configuration
+@Configuration
 @EnableJpaRepositories(basePackages = "com.businessassistantbcn.mydata.repository")
 @EntityScan(basePackages = "com.businessassistantbcn.mydata.entities")
 @PropertySource("classpath:application-test.properties")
-@EnableTransactionManagement*/
-public class SpringDBTestConfiguration {
+@EnableTransactionManagement
+public class DBTestConfiguration {
 
-/*    @Autowired
+    @Autowired
     private Environment env;
 
     @Bean
@@ -39,5 +31,5 @@ public class SpringDBTestConfiguration {
         dataSourceBuilder.username(env.getProperty("spring.datasource.user"));
         dataSourceBuilder.password(env.getProperty("spring.datasource.password"));
         return dataSourceBuilder.build();
-    }*/
+    }
 }
