+++ BusinessAssistant-mydata/src/test/resources/application-test.properties
@@ -18,5 +18,3 @@ spring.sql.init.mode=always
 logging.level.org.hibernate.SQL=INFO
 logging.level.org.hibernate.type=TRACE
 
-
-#spring.profiles.active=test
